Add Shape and Fluid objects to Solver class"
See: #1
