+++ utils/fluids.py
@@ -4,3 +4,4 @@ from dataclasses import dataclass
 @dataclass
 class Fluid:
     p: float = 1
+    name: str = "He"
